Checksums tests (#3826)
