+++ catalog/package.json
@@ -152,6 +152,7 @@
     "@graphql-codegen/typescript-operations": "^1.18.4",
     "@graphql-codegen/urql-introspection": "^1.3.1",
     "@graphql-typed-document-node/core": "^3.2.0",
+    "@peculiar/webcrypto": "^1.4.3",
     "@testing-library/react": "^12.1.5",
     "@testing-library/react-hooks": "^8.0.1",
     "@types/d3-scale": "^4.0.8",
@@ -183,6 +184,7 @@
     "@types/zxcvbn": "^4.4.1",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
     "@typescript-eslint/parser": "^6.3.0",
+    "blob-polyfill": "^7.0.20220408",
     "bundlewatch": "^0.3.3",
     "chalk": "^4.1.2",
     "circular-dependency-plugin": "^5.2.2",
